some serliazation with sideloading didn't want to make too many decision b/c not totally sure how these endpoint will interact with React but so far: cohort side loads all its students but not its schedule, schedule side loads all attendance records, organized by day, but only for days that are less than or equal to today, attendance record side loads its associated student.
Need to think about way the user will interact. Probably something like:
visit Cohort show page and see schedule (as calendar) -> click on a specific day to see details of attendance for that day.
Use should also be able to view all attendance records for a given student, look at a bunch of days and all their records as once